refactor(aviation): extract API response types into types.rs
